requirements.txt
celery
django-celery
django-celery-email